requirements.txt
flask==3.0.3
openai==1.52.2
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2